refacto: move param parser into the class they are building
