+++ search-service/src/main/kotlin/com/egm/stellio/search/entity/model/EntitiesQuery.kt
@@ -3,9 +3,9 @@ package com.egm.stellio.search.entity.model
 import com.egm.stellio.shared.model.EntitySelector
 import com.egm.stellio.shared.model.EntityTypeSelection
 import com.egm.stellio.shared.model.ExpandedTerm
-import com.egm.stellio.shared.model.parameter.GeoQuery
 import com.egm.stellio.shared.model.LinkedEntityQuery
 import com.egm.stellio.shared.model.PaginationQuery
+import com.egm.stellio.shared.model.parameter.GeoQuery
 import java.net.URI
 
 sealed class EntitiesQuery(
