+++ search-service/src/main/kotlin/com/egm/stellio/search/entity/util/EntitiesQueryUtils.kt
@@ -10,14 +10,14 @@ import com.egm.stellio.shared.config.ApplicationProperties
 import com.egm.stellio.shared.model.APIException
 import com.egm.stellio.shared.model.BadRequestDataException
 import com.egm.stellio.shared.model.EntitySelector
+import com.egm.stellio.shared.model.LinkedEntityQuery.Companion.parseLinkedEntityQueryParameters
+import com.egm.stellio.shared.model.PaginationQuery.Companion.parsePaginationParameters
 import com.egm.stellio.shared.model.parameter.GeoQueryParameter.Companion.parseGeoQueryParameters
 import com.egm.stellio.shared.model.parameter.QueryParam
 import com.egm.stellio.shared.util.JsonLdUtils
 import com.egm.stellio.shared.util.decode
 import com.egm.stellio.shared.util.expandTypeSelection
 import com.egm.stellio.shared.util.parseAndExpandRequestParameter
-import com.egm.stellio.shared.util.parseLinkedEntityQueryParameters
-import com.egm.stellio.shared.util.parsePaginationParameters
 import com.egm.stellio.shared.util.parseRequestParameter
 import com.egm.stellio.shared.util.toListOfUri
 import com.egm.stellio.shared.util.validateIdPattern
