+++ shared/src/main/kotlin/com/egm/stellio/shared/model/PaginationQuery.kt
@@ -14,14 +14,15 @@ data class PaginationQuery(
     val count: Boolean = false
 ) {
     companion object {
-        fun create(
-            params: MultiValueMap<String, String>,
+
+        fun parsePaginationParameters(
+            queryParams: MultiValueMap<String, String>,
             limitDefault: Int,
             limitMax: Int
         ): Either<APIException, PaginationQuery> {
-            val count = params.getFirst(COUNT.key)?.toBoolean() ?: false
-            val offset = params.getFirst(OFFSET.key)?.toIntOrNull() ?: 0
-            val limit = params.getFirst(LIMIT.key)?.toIntOrNull() ?: limitDefault
+            val count = queryParams.getFirst(COUNT.key)?.toBoolean() ?: false
+            val offset = queryParams.getFirst(OFFSET.key)?.toIntOrNull() ?: 0
+            val limit = queryParams.getFirst(LIMIT.key)?.toIntOrNull() ?: limitDefault
             if (!count && (limit <= 0 || offset < 0))
                 return BadRequestDataException(
                     "Offset must be greater than zero and limit must be strictly greater than zero"
